Refactor services test suite to use server config
